+++ CHANGELOG.rst
@@ -22,6 +22,7 @@ Fixed:
 - Enforce camelCaseFields and Alphanumeric + underscore + dash Part names
 - Don't load child runnable block designs at init
 - Error message when a Field is not writeable now shows current state
+- Check generator units match axis units in pmac cs part
 
 `3-0a3`_ - 2018-07-25
 ---------------------